third_party/WebKit/LayoutTests/inspector/elements/styles/style-autocomplete-expected.txt

Tests that autocompletions are computed correctly when editing the Styles pane.


Running: testEmptyName

Running: testEmptyNameForce

Running: testSingleCharName

Running: testEmptyValue

Running: testValueR

Running: testValueWithParenthesis

Running: testValuePrefixed

Running: testValueUnprefixed
